+++ package.json
@@ -10,9 +10,9 @@
 		"preview": "vite preview"
 	},
 	"dependencies": {
-		"chroma-js": "^3.1.2",
 		"axios": "^1.10.0",
 		"chart.js": "^4.5.0",
+		"chroma-js": "^3.1.2",
 		"react": "^19.0.0",
 		"react-calendar": "^5.1.0",
 		"react-chartjs-2": "^5.3.0",
@@ -28,7 +28,7 @@
 		"@eslint/js": "^9.22.0",
 		"@types/react": "^19.0.10",
 		"@types/react-dom": "^19.0.4",
-		"@vitejs/plugin-react": "^4.3.4",
+		"@vitejs/plugin-react": "^4.6.0",
 		"eslint": "^9.26.0",
 		"eslint-plugin-jsx-a11y": "^6.10.2",
 		"eslint-plugin-no-relative-import-paths": "^1.6.1",
@@ -38,7 +38,7 @@
 		"eslint-plugin-react-refresh": "^0.4.19",
 		"globals": "^16.0.0",
 		"prettier": "^3.5.3",
-		"vite": "^7.0.0",
+		"vite": "^6.3.5",
 		"vite-plugin-svgr": "^4.3.0"
 	}
 }
