+++ src/pages/admin/AdminPlays.jsx
@@ -1,11 +1,11 @@
 import styled from 'styled-components';
 import UserTable from '@/components/Admin/UserTable';
-import Search from '@/assets/icons/searchBlack.svg?react';
-import SearchBg from '@/assets/icons/searchBlackBg.svg?react';
+import FilterHeader from '@/components/Admin/FilterHeader';
+
 import { useState, useMemo } from 'react';
 
 function AdminPlays() {
-    //상태 별 글자 색상 변경 추가 필요
+	//상태 별 글자 색상 변경 추가 필요
 	const play_data = [
 		{
 			title: '소극장 공연 이름',
@@ -68,12 +68,11 @@ function AdminPlays() {
 	const [currentPage, setCurrentPage] = useState(1);
 	const itemsPerPage = 20;
 
-	const handleColumnToggle = (column) => {
-		setVisibleColumns((prev) =>
-			prev.includes(column)
-				? prev.filter((c) => c !== column)
-				: [...prev, column],
-		);
+	const filterKeys = ['title', 'date', 'uploader'];
+	const filterLabels = {
+		title: '공연 이름',
+		date: '날짜/시간',
+		uploader: '등록자명',
 	};
 
 	const filteredData = useMemo(() => {
@@ -96,40 +95,17 @@ function AdminPlays() {
 
 	return (
 		<Container>
-			<Top>
-				<Search />
-			</Top>
 			<Content>
-				<Sidebar />
 				<TableArea>
-					<Title>사용자 관리</Title>
-					<FilterArea>
-						<SearchInput>
-							<input
-								type="text"
-								value={searchTerm}
-								onChange={(e) => {
-									setSearchTerm(e.target.value);
-									setCurrentPage(1);
-								}}
-								placeholder="검색어를 입력하세요"
-							/>
-							<Search width={15} />
-						</SearchInput>
-						<div className="checkboxArea">
-							{['title', 'date', 'uploader'].map((key) => (
-								<label key={key}>
-									<input
-										type="checkbox"
-										checked={visibleColumns.includes(key)}
-										onChange={() => handleColumnToggle(key)}
-									/>
-									{play_data[0][key]}
-								</label>
-							))}
-							<SearchBg />
-						</div>
-					</FilterArea>
+					<FilterHeader
+						title="소극장 공연 관리"
+						searchTerm={searchTerm}
+						setSearchTerm={setSearchTerm}
+						filterKeys={filterKeys}
+						filterLabels={filterLabels}
+						visibleColumns={visibleColumns}
+						setVisibleColumns={setVisibleColumns}
+					/>
 
 					<UserTable
 						data={[play_data[0], ...paginatedData]}
@@ -152,64 +128,11 @@ const Container = styled.div`
 	display: flex;
 	flex-direction: column;
 `;
-const Top = styled.div`
-	height: 108px;
-	display: flex;
-	justify-content: flex-end;
-	padding: 40px 110px;
-`;
-
 const Content = styled.div`
 	width: 100%;
 	display: flex;
 `;
-const Sidebar = styled.div`
-	//추후 컴포넌트로 변경 후 삭제
-	width: 290px;
-	height: 100vh;
-	position: fixed;
-	padding: 27px 18px;
-	background: #8f8e94;
-`;
-const FilterArea = styled.div`
-	display: flex;
-	justify-content: space-between;
-	margin-bottom: 36px;
-
-	.checkboxArea {
-		display: flex;
-		align-items: center;
-		gap: 18px;
-	}
-`;
-const SearchInput = styled.div`
-	display: flex;
-	align-items: center;
-	padding: 0 10px;
-	background: #fff;
-	width: 360px;
-	height: 32px;
-	border-radius: 7px;
-	border: 1px solid #000;
-
-	input {
-		width: 100%;
-		border: none;
-		outline: none;
-		font-size: ${({ theme }) => theme.font.fontSize.body14};
-		font-weight: ${({ theme }) => theme.font.fontWeight.bold};
-		color: ${({ theme }) => theme.colors.grayMain};
-	}
-`;
-
 const TableArea = styled.div`
-	margin-left: 290px;
 	padding: 0px 120px 50px 50px;
 	width: 100%;
-`;
-const Title = styled.h3`
-	font-size: ${({ theme }) => theme.font.fontSize.headline24};
-	font-weight: ${({ theme }) => theme.font.fontWeight.bold};
-	color: ${({ theme }) => theme.colors.pink600};
-	margin-bottom: 15px;
 `;
