+++ src/pages/admin/GalleryDetail.jsx
@@ -8,11 +8,7 @@ function GalleryDetail() {
 	const [searchTerm, setSearchTerm] = useState('');
 	return (
 		<Container>
-			<Top>
-				<Search />
-			</Top>
 			<Content>
-				<Sidebar />
 				<GallData>
 					<Title>사진첩 관리</Title>
 					<div className="serachNadd">
@@ -50,19 +46,12 @@ export default GalleryDetail;
 const Container = styled.div`
 	width: 100vw;
 `;
-const Top = styled.div`
-	height: 108px;
-	display: flex;
-	justify-content: flex-end;
-	padding: 40px 110px;
-`;
 
 const Content = styled.div`
 	width: 100%;
 	display: flex;
 `;
 const GallData = styled.div`
-	margin-left: 290px;
 	width: 100%;
 	padding: 0px 120px 50px 50px;
 
@@ -94,14 +83,6 @@ const GallData = styled.div`
 	}
 `;
 
-const Sidebar = styled.div`
-	//추후 컴포넌트로 변경 후 삭제
-	width: 290px;
-	height: 100vh;
-	position: fixed;
-	padding: 27px 18px;
-	background: #8f8e94;
-`;
 const Title = styled.h3`
 	font-size: ${({ theme }) => theme.font.fontSize.headline24};
 	font-weight: ${({ theme }) => theme.font.fontWeight.bold};
