+++ src/pages/gallery/Gallery.jsx
@@ -10,7 +10,6 @@ import image3 from '@/assets/mock/images/image3.png';
 import image4 from '@/assets/mock/images/image4.png';
 import image5 from '@/assets/mock/images/image5.png';
 import styled from 'styled-components';
-import HomeIconMenu from '../../components/HomeIconMenu';
 
 function Gallery() {
 	const imageList = [
@@ -43,7 +42,6 @@ function Gallery() {
 			</Mobile>
 
 			<Web>
-
 				<SideMenuWrapper>
 					<HomeIconMenu isWeb={true} />
 				</SideMenuWrapper>
@@ -90,19 +88,6 @@ const Mobile = styled.div`
 	}
 `;
 
-const SideMenuWrapper = styled.div`
-	width: 101px;
-	height: 100vh;
-	position: fixed;
-	top: 0;
-	left: 0;
-	flex-shrink: 0;
-	display: none;
-	background-color: white;
-	@media (min-width: 768px) {
-		display: block;
-	}
-`;
 const Container = styled.div`
 	width: 100%;
 	padding: 60px 100px 100px 60px;
